+++ CHANGES.md
@@ -1,3 +1,12 @@
+0.2.0
+-----
+
+- Make sure to import command line arguments registered with
+  ocaml-migrate-parsetree
+
+- Fix an issue where cookies set from the command line sometimes
+  disappeared
+
 0.1.0
 -----